Make CollapseEmptyPlan.scala.
